requirements.txt
streamlit==0.64.0
pandas==1.0.5
Pillow==7.2.0
scikit_learn==0.23.1
